# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orders API Documentation",
        "description": "Documentação da API de gestão de encomendas e produtos",
        "version": "1.0.0"
    },
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Listar todas as encomendas",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Criar uma nova encomenda",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_name",
                                    "customer_email",
                                    "products"
                                ],
                                "properties": {
                                    "customer_name": {
                                        "type": "string",
                                        "example": "João Silva"
                                    },
                                    "customer_email": {
                                        "type": "string",
                                        "example": "joao@email.com"
                                    },
                                    "products": {
                                        "description": "Lista de produtos da encomenda",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Encomenda criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação ou stock insuficiente"
                    }
                }
            }
        },
        "/api/orders/{order}/cancel": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancelar uma encomenda",
                "operationId": "6b3a53e3f3f98b4e5da616fa85e4f37d",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID da encomenda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Encomenda cancelada com sucesso"
                    },
                    "422": {
                        "description": "Encomenda já se encontra cancelada"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar todos os produtos",
                "operationId": "d97d1c3727cbac49708b85f7d11f7d58",
                "responses": {
                    "200": {
                        "description": "Sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Criar um novo produto",
                "operationId": "785644b7d7dd38af4b51dafd0fbc1b6d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Teclado Mecânico"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 79.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/products/{product}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualizar um produto existente",
                "operationId": "f47cf55f4f3e0b134b5e7d9a8373725d",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Teclado Mecânico RGB"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 89.99
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 20
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Eliminar um produto",
                "operationId": "ac878db81fdb1333b21fc79e574a5a17",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto eliminado com sucesso"
                    },
                    "404": {
                        "description": "Produto não encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Orders",
            "description": "Orders"
        },
        {
            "name": "Products",
            "description": "Products"
        }
    ]
}